Auto merge partitioned netcdf files (#5)
The major change in this PR is the introduction of mppnccombine tool into the FMS diag_manager. When an IO_LAYOUT is specified, FMS now automatically merges the partitioned files when diag manager closes the open netcdf files. For domain files, this is done in a round robin fashion, where each IO PE takes care of the auto-merging of a history file stream. For section files, the IO PE with the smallest PE index takes care of the merging. The auto merge feature is enabled by the newly added auto_merge_nc namelist parameter.


* first working version of auto nccombine execution

* reme existing outfile, if an old version exists

* first complete version of automated file combiner

* set extent_type to 1 for fileobjects opened with open_domain_file

* use select(type) block to determine netcdf file type (domain or not) and remove existing combined file when root pe opens a partitioned one.

* introduce auto_merge_nc nml parameter and simplify mppnccombine function interfaces

* split the identification of files to combine and execution of mppnccombine via a barrier

* fix round robin combination algm

by determining which IO PE combines which file based
on the order of IO PES (as opposed to IO PE index,
which is not guaranteed to be contiguous when
land block elimination is enabled)

* comment out Combining file log